% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{addition}
\alias{addition}
\title{Additionne deux variables}
\usage{
addition(nb1, nb2)
}
\arguments{
\item{nb1}{un nombre}

\item{nb2}{un nombre}
}
\value{
la fonction renvoie la somme des deux parametres
}
\description{
Additionne deux variables
}
\examples{
addition(5, 10)

}
